feat(ui): implement reusable FilesPage and associated components
