+++ app/(logged_in)/publications/page.tsx
@@ -3,7 +3,6 @@ import { Box } from '@mui/material';
 export default function PublicationsPage() {
   return (
     <Box>
-
     </Box>
   );
 }
